little tweaks, rephrased questions, etc.
rephrased the questions so they actually seem like questions instead of a simple list of issues. this is an FAQ after all. there is more room for improvement though. i'd say that the top line is still unnecessary because there is no contact to Juli listed anywhere in the repo, a link to the discord server would suffice i'd say that the last issue, the one with the viewmodel, should be more clear. someone new might not know what these values do, let alone changing them correctly.